Split JS marks into separate files
